[wip] first pass on converting notifier over to using middleware with a “legacy” middleware
