+++ lib/slack-notifier/payload_middleware.rb
@@ -18,3 +18,4 @@ end
 
 require_relative 'payload_middleware/stack'
 require_relative 'payload_middleware/base'
+require_relative 'payload_middleware/legacy'
